Restructure portfolio into composable server-first components
- Split the Workshop portfolio into per-section files under
  src/components/portfolio/ (Hero, WorkGrid, Publications, TalksList,
  Awards, WritingList, NowPlaying, Contact, Toast, SystemAudit,
  Flourishes, Shared, SocialIcons, ClientChrome, hooks, data, lifeline)
- Make every static section an async server component so editorial
  copy ships in the initial HTML for SEO; ClientChrome handles the
  sticky Nav, scroll-spy, keyboard easter eggs, toast and audit overlays
- Wire i18n via next-intl rich text: t.rich(key, richTags) with shared
  em/strong renderers
- Re-add i18n locale files (en-US, pt-BR) keyed to the new design copy
- Calculate years-in-industry, years-tinkering and the current-month
  label dynamically from a small lifeline.ts so the bio stays fresh
- Add /api/coin route handler (Yahoo Finance, 5-min ISR cache) and use
  it for the real $COIN ticker on the Coinbase tile
- Remove the standalone `b` shortcut (collided with the Konami suffix);
  Konami still triggers CRT scanlines, toast and beer rain
- Redesign /[locale]/links with the new design system
- Delete legacy components (sections/, illustrations/, theme-toggle,
  language-select, work-experience, consts, scripts, env.ts) and the
  old single-file portfolio.tsx
