--- src/@types/i18n.d.ts
@@ -1,3 +1,0 @@
-import type Messages from "../i18n/locales/en-US.json";
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-declare interface IntlMessages extends Messages {}
