+++ src/app/[locale]/layout.tsx
@@ -1,22 +1,27 @@
 import { NextIntlClientProvider } from "next-intl";
 import { getMessages } from "next-intl/server";
+import { notFound } from "next/navigation";
+import { routing } from "@/utils/routing";
 
-const LocaleLayout = async ({
+export function generateStaticParams() {
+  return routing.locales.map((locale) => ({ locale }));
+}
+
+export default async function LocaleLayout({
   children,
   params,
-}: React.PropsWithChildren<{
+}: {
+  children: React.ReactNode;
   params: Promise<{ locale: string }>;
-}>) => {
+}) {
   const { locale } = await params;
-  // Providing all messages to the client
-  // side is the easiest way to get started
+  if (!routing.locales.includes(locale as (typeof routing.locales)[number])) {
+    notFound();
+  }
   const messages = await getMessages();
-
   return (
     <NextIntlClientProvider messages={messages} locale={locale}>
       {children}
     </NextIntlClientProvider>
   );
-};
-
-export default LocaleLayout;
+}
