+++ src/app/[locale]/page.tsx
@@ -1,4 +1,4 @@
-import Portfolio from "./portfolio";
+import { Portfolio } from "@/components/portfolio";
 
 export default function Page() {
   return <Portfolio />;
