--- src/components/illustrations/awards-illustration.tsx
@@ -1,16 +1,0 @@
-export const AwardsIllustration = () => (
-  <div className="flex items-end absolute -translate-y-20 -z-10">
-    <div>
-      <div className="w-28 md:w-36 h-6 from-black/5 dark:bg-zinc-900" />
-      <div className="w-28 md:w-36 h-32 from-black/5 dark:from-zinc-900 to-transparent bg-linear-to-b" />
-    </div>
-    <div>
-      <div className="w-28 md:w-36 h-24 bg-black/10 dark:bg-zinc-800" />
-      <div className="w-28 md:w-36 h-32 from-black/10 dark:from-zinc-800 to-transparent bg-linear-to-b" />
-    </div>
-    <div>
-      <div className="w-28 md:w-36 h-12 bg-black/5 dark:bg-zinc-900" />
-      <div className="w-28 md:w-36 h-32 from-black/5 dark:from-zinc-900 to-transparent bg-linear-to-b" />
-    </div>
-  </div>
-);
