--- src/components/illustrations/location-icon.tsx
@@ -1,7 +1,0 @@
-import type { SVGProps } from "react";
-
-export const LocationIcon = (props: SVGProps<SVGSVGElement>) => (
-  <svg xmlns="http://www.w3.org/2000/svg" width={18} height={17} {...props}>
-    <path d="M9 14.875a21.085 21.085 0 0 1-2.479-2.501c-1.133-1.354-2.479-3.37-2.479-5.29a4.958 4.958 0 0 1 8.466-3.508 4.926 4.926 0 0 1 1.45 3.507c0 1.921-1.345 3.937-2.479 5.29A21.085 21.085 0 0 1 9 14.876Zm0-9.917a2.125 2.125 0 1 0 0 4.25 2.125 2.125 0 0 0 0-4.25Z" />
-  </svg>
-);
