--- src/components/sections/index.ts
@@ -1,6 +1,0 @@
-export { Awards } from "./awards";
-export { Hero } from "./hero";
-export { Publications } from "./publications";
-export { Socials } from "./socials";
-export { Summary } from "./summary";
-export { WorkExperiences } from "./work-experiences";
