--- src/consts/videos.json
@@ -1,147 +1,0 @@
-[
-  {
-    "name": "Sistemas Legados e Organização com Gi Bordignon | Space Cast #40",
-    "url": "https://www.youtube.com/watch?v=FD2BsDW7jJg",
-    "image": "https://i.ytimg.com/vi/FD2BsDW7jJg/maxresdefault.jpg"
-  },
-  {
-    "name": "Ser DevRel no Github com Pachi Parra | Space Cast #39",
-    "url": "https://www.youtube.com/watch?v=D6QtBuJoQ1A",
-    "image": "https://i.ytimg.com/vi/D6QtBuJoQ1A/maxresdefault.jpg"
-  },
-  {
-    "name": "Se Tornando Dev com Simara Conceição | Space Cast #38",
-    "url": "https://www.youtube.com/watch?v=r_g_UpP5j1o",
-    "image": "https://i.ytimg.com/vi/r_g_UpP5j1o/maxresdefault.jpg"
-  },
-  {
-    "name": "Mercado Dev para 40+ com Dan Vitoriano | Space Cast #37",
-    "url": "https://www.youtube.com/watch?v=UIYb8zd02ZY",
-    "image": "https://i.ytimg.com/vi/UIYb8zd02ZY/maxresdefault.jpg"
-  },
-  {
-    "name": "Abraçando o Caos com Josi Assis | Space Cast #36",
-    "url": "https://www.youtube.com/watch?v=DLEaut5emsU",
-    "image": "https://i.ytimg.com/vi/DLEaut5emsU/maxresdefault.jpg"
-  },
-  {
-    "name": "Decolando sua Carreira Júnior com Fernanda Kipper | Space Cast #35",
-    "url": "https://www.youtube.com/watch?v=FAzEPoQBY2I",
-    "image": "https://i.ytimg.com/vi/FAzEPoQBY2I/maxresdefault.jpg"
-  },
-  {
-    "name": "Mercado de Tecnologia para Pessoas Atípicas com Gui Estevão | Space Cast #34",
-    "url": "https://www.youtube.com/watch?v=ErHBAMZ9qjI",
-    "image": "https://i.ytimg.com/vi/ErHBAMZ9qjI/maxresdefault.jpg"
-  },
-  {
-    "name": "Transformação Digital Em Empresas Centenárias com Heitor Penha | Space Cast #33",
-    "url": "https://www.youtube.com/watch?v=JLMvXHPZABE",
-    "image": "https://i.ytimg.com/vi/JLMvXHPZABE/maxresdefault.jpg"
-  },
-  {
-    "name": "Liderança Ágil com Alexandre Abdalla | Space Cast #32",
-    "url": "https://www.youtube.com/watch?v=ioQ5gjjOaf4",
-    "image": "https://i.ytimg.com/vi/ioQ5gjjOaf4/maxresdefault.jpg"
-  },
-  {
-    "name": "Construindo A Sua História com Joel Backschat | Space Cast #31",
-    "url": "https://www.youtube.com/watch?v=3DZEftUeSLM",
-    "image": "https://i.ytimg.com/vi/3DZEftUeSLM/maxresdefault.jpg"
-  },
-  {
-    "name": "Desenvolvimento Mobile com Rodrigo Gonçalves | Space Cast #30",
-    "url": "https://www.youtube.com/watch?v=n7ppFE9xybw",
-    "image": "https://i.ytimg.com/vi/n7ppFE9xybw/maxresdefault.jpg"
-  },
-  {
-    "name": "Como ser Fullcycle com Wesley Willians | Space Cast #29",
-    "url": "https://www.youtube.com/watch?v=hITHDVvTYnk",
-    "image": "https://i.ytimg.com/vi/hITHDVvTYnk/maxresdefault.jpg"
-  },
-  {
-    "name": "Perspectiva Gringa: Devs Brasileiros com Patrick Cason | Space Cast #28",
-    "url": "https://www.youtube.com/watch?v=kD0kf75N7WE",
-    "image": "https://i.ytimg.com/vi/kD0kf75N7WE/maxresdefault.jpg"
-  },
-  {
-    "name": "Conheça o Space Squad com Julião | Space Cast #27",
-    "url": "https://www.youtube.com/watch?v=jtyi0PmJzLQ",
-    "image": "https://i.ytimg.com/vi/jtyi0PmJzLQ/maxresdefault.jpg"
-  },
-  {
-    "name": "Carreira de DevRel com Gabriel Bolzi | Space Cast #26",
-    "url": "https://www.youtube.com/watch?v=tKc91BExzrY",
-    "image": "https://i.ytimg.com/vi/tKc91BExzrY/maxresdefault.jpg"
-  },
-  {
-    "name": "O Mundo da Inteligência Artificial com Carla Vieira | Space Cast #25",
-    "url": "https://www.youtube.com/watch?v=lX2v1TS7gM4",
-    "image": "https://i.ytimg.com/vi/lX2v1TS7gM4/maxresdefault.jpg"
-  },
-  {
-    "name": "Cybersegurança com Laís e Lillian | Space Cast #24",
-    "url": "https://www.youtube.com/watch?v=I3UykAUPvAs",
-    "image": "https://i.ytimg.com/vi/I3UykAUPvAs/maxresdefault.jpg"
-  },
-  {
-    "name": "Comunidade PHP com PokemãoBR | Space Cast #23",
-    "url": "https://www.youtube.com/watch?v=Y0lL018lwAs",
-    "image": "https://i.ytimg.com/vi/Y0lL018lwAs/maxresdefault.jpg"
-  },
-  {
-    "name": "Redes Sociais para Devs com Valks Carvalho | Space Cast #22",
-    "url": "https://www.youtube.com/watch?v=fIbMAq43OFs",
-    "image": "https://i.ytimg.com/vi/fIbMAq43OFs/maxresdefault.jpg"
-  },
-  {
-    "name": "Encontrando Um Propósito com Abraão Sena | Space Cast #21",
-    "url": "https://www.youtube.com/watch?v=Gd1DpDf65HM",
-    "image": "https://i.ytimg.com/vi/Gd1DpDf65HM/maxresdefault.jpg"
-  },
-  {
-    "name": "Evolução Constante com Kamila Code | Space Cast #20",
-    "url": "https://www.youtube.com/watch?v=bgTW6PInZDs",
-    "image": "https://i.ytimg.com/vi/bgTW6PInZDs/maxresdefault.jpg"
-  },
-  {
-    "name": "Agarrando Oportunidades com Vitor Castro | Space Cast #19",
-    "url": "https://www.youtube.com/watch?v=mDWqnYIgUy0",
-    "image": "https://i.ytimg.com/vi/mDWqnYIgUy0/maxresdefault.jpg"
-  },
-  {
-    "name": "Angular no Exterior com Leonardo Natal | Space Cast #18",
-    "url": "https://www.youtube.com/watch?v=MNeH30-zr1o",
-    "image": "https://i.ytimg.com/vi/MNeH30-zr1o/maxresdefault.jpg"
-  },
-  {
-    "name": "A Interface Perfeita com Jonas Milan | Space Cast #17",
-    "url": "https://www.youtube.com/watch?v=H6O5vxtXRfM",
-    "image": "https://i.ytimg.com/vi/H6O5vxtXRfM/maxresdefault.jpg"
-  },
-  {
-    "name": "Trabalhando No Exterior com Ilda Neta | Space Cast #16",
-    "url": "https://www.youtube.com/watch?v=oR8E5yTGIo4",
-    "image": "https://i.ytimg.com/vi/oR8E5yTGIo4/maxresdefault.jpg"
-  },
-  {
-    "name": "A Pressão do SHOW ME THE CODE com José Carlos Teles | Space Cast #15",
-    "url": "https://www.youtube.com/watch?v=xmhFe-7NvcI",
-    "image": "https://i.ytimg.com/vi/xmhFe-7NvcI/maxresdefault.jpg"
-  },
-  {
-    "name": "Equilibrando vida pessoal e trabalho com Laura Kibum | Space Cast #14",
-    "url": "https://www.youtube.com/watch?v=mH42lps6O7A",
-    "image": "https://i.ytimg.com/vi/mH42lps6O7A/maxresdefault.jpg"
-  },
-  {
-    "name": "Qual o Papel de um Head de Engenharia com Sergio Fiorotti | Space Cast #13",
-    "url": "https://www.youtube.com/watch?v=4rlktCGOFRQ",
-    "image": "https://i.ytimg.com/vi/4rlktCGOFRQ/maxresdefault.jpg"
-  },
-  {
-    "name": "Nathally Souza | Space Cast #12",
-    "url": "https://www.youtube.com/watch?v=mCPU7G9IqCA",
-    "image": "https://i.ytimg.com/vi/mCPU7G9IqCA/maxresdefault.jpg"
-  }
-]
